backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
pypdf==4.3.1
python-docx==1.1.2
openpyxl==3.1.5
python-pptx==1.0.2
Pillow==10.4.0
reportlab==4.2.2
markdown==3.7
pdfkit==1.0.0
aiofiles==24.1.0
